refactor: domain에 따라 패키지 구조 변경
* memo, page, user, common에 따라 크게 패키지 구조를 나눴습니다
* 공통적으로 처리되는 dto, exception은 common 안으로 옮겼습니다
* 3tier architecture에 맞게 application, controller, repository로
  나누었습니다
  * dto는 service에 종속된다고 생각하여 application에 넣었습니다
  * entity는 jpa repository에 종속된다 생각하여 repository에 넣었습니다
