+++ src/main/java/dev/gooiman/server/common/exception/BaseException.java
@@ -1,0 +1,9 @@
+package dev.gooiman.server.common.exception;
+
+import lombok.Getter;
+
+@Getter
+public class BaseException extends Exception {
+
+    private BaseResponseStatus baseResponseStatus;
+}
