--- src/main/java/dev/gooiman/server/exception/BaseException.java
@@ -1,9 +1,0 @@
-package dev.gooiman.server.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class BaseException extends Exception{
-    private BaseResponseStatus baseResponseStatus;
-}
