+++ src/main/java/dev/gooiman/server/memo/application/dto/CreateMemoResponseDto.java
@@ -1,13 +1,13 @@
-package dev.gooiman.server.dto;
+package dev.gooiman.server.memo.application.dto;
 
-import dev.gooiman.server.domain.Memo;
+import dev.gooiman.server.memo.repository.entity.Memo;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class CreateMemoDto {
+public class CreateMemoResponseDto {
 
     private String pageId;
     private String author;
@@ -18,7 +18,8 @@ public class CreateMemoDto {
 
     @Getter
     @Builder
-    public static class Res{
+    public static class Res {
+
         private String pageId;
         private String author;
         private String title;
@@ -28,12 +29,12 @@ public class CreateMemoDto {
 
         public static Res mapEntityToDto(Memo memoEntity) {
             return Res.builder().pageId(String.valueOf(memoEntity.getPage().getPageId()))
-                    .author(memoEntity.getUser().getName())
-                    .title(memoEntity.getTitle())
-                    .category(memoEntity.getCategory())
-                    .subCategory(memoEntity.getSubCategory())
-                    .content(memoEntity.getContent())
-                    .build();
+                .author(memoEntity.getUser().getName())
+                .title(memoEntity.getTitle())
+                .category(memoEntity.getCategory())
+                .subCategory(memoEntity.getSubCategory())
+                .content(memoEntity.getContent())
+                .build();
         }
     }
