+++ src/main/java/dev/gooiman/server/memo/repository/entity/Memo.java
@@ -1,13 +1,20 @@
-package dev.gooiman.server.domain;
+package dev.gooiman.server.memo.repository.entity;
 
-import jakarta.persistence.*;
+import dev.gooiman.server.page.repository.entity.Page;
+import dev.gooiman.server.user.repository.entity.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.GenericGenerator;
-
-import java.util.UUID;
 
 @Getter
 @NoArgsConstructor
