+++ src/main/java/dev/gooiman/server/page/application/PageService.java
@@ -1,15 +1,13 @@
-package dev.gooiman.server.service;
+package dev.gooiman.server.page.application;
 
-import dev.gooiman.server.domain.Memo;
-import dev.gooiman.server.domain.Page;
-import dev.gooiman.server.dto.CreatePageDto;
-import dev.gooiman.server.dto.MemoSummariseDto;
-import dev.gooiman.server.repository.MemoRepository;
-import dev.gooiman.server.repository.PageRepository;
+import dev.gooiman.server.page.repository.entity.Page;
+import dev.gooiman.server.page.application.dto.CreatePageResponseDto;
+import dev.gooiman.server.memo.application.dto.MemoSummariseResponseDto;
+import dev.gooiman.server.memo.repository.MemoRepository;
+import dev.gooiman.server.page.repository.PageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -20,14 +18,14 @@ public class PageService {
     private final PageRepository pageRepository;
     private final MemoRepository memoRepository;
 
-    public CreatePageDto.Res create(CreatePageDto createPageDto) {
+    public CreatePageResponseDto.Res create(CreatePageResponseDto createPageDto) {
         String name = createPageDto.getName();
         Page page = Page.builder().pageName(name).build();
         Page savedPage = pageRepository.save(page);
-        return CreatePageDto.Res.mapEntityToDto(savedPage);
+        return CreatePageResponseDto.Res.mapEntityToDto(savedPage);
     }
 
-    public MemoSummariseDto.Res memoSummarise(String pageId) {
+    public MemoSummariseResponseDto.Res memoSummarise(String pageId) {
         Optional<Page> page = pageRepository.findById(UUID.fromString(pageId));
         Page pageEntity = page.get();
         String pageName = pageEntity.getPageName();
