+++ src/main/java/dev/gooiman/server/page/application/dto/CreatePageResponseDto.java
@@ -1,17 +1,19 @@
-package dev.gooiman.server.dto;
+package dev.gooiman.server.page.application.dto;
 
-import dev.gooiman.server.domain.Page;
+import dev.gooiman.server.page.repository.entity.Page;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class CreatePageDto {
+public class CreatePageResponseDto {
+
     private String name;
 
     @Getter
     @AllArgsConstructor
     public static class Res {
+
         private String id;
 
         public static Res mapEntityToDto(Page pageEntity) {
