--- src/main/java/dev/gooiman/server/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package dev.gooiman.server.repository;
-
-import dev.gooiman.server.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, UUID> {
-
-    Optional<User> findByPageAndName(String pageId, String name);
-}
