+++ src/main/java/dev/gooiman/server/user/repository/entity/User.java
@@ -1,14 +1,19 @@
-package dev.gooiman.server.domain;
+package dev.gooiman.server.user.repository.entity;
 
-import jakarta.persistence.*;
+import dev.gooiman.server.page.repository.entity.Page;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.GenericGenerator;
-
-import java.util.List;
-import java.util.UUID;
 
 @Getter
 @NoArgsConstructor
@@ -17,6 +22,7 @@ import java.util.UUID;
 @Table(name = "USERS")
 @Entity
 public class User {
+
     @Id
     @GeneratedValue(generator = "uuid2")
     @GenericGenerator(name = "uuid2", strategy = "uuid2")
